+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rust-ssr-client"
+name = "ssr-client"
 version = "0.1.0"
 edition = "2024"
